Add supprt for WhatToMine.com

+++ MultiMiner.CoinApi/Data/CoinInformation.cs
@@ -7,11 +7,11 @@ namespace MultiMiner.CoinApi.Data
         public string Symbol { get; set; }
         public string Name { get; set; }
         public string Algorithm { get; set; }
-        public int CurrentBlocks { get; set; }
+        public long CurrentBlocks { get; set; }
         public double Difficulty { get; set; }
         public double Reward { get; set; }
         public double MinimumBlockTime { get; set; }
-        public Int64 NetworkHashRate { get; set; }
+        public double NetworkHashRate { get; set; }
         public double Price { get; set; }
         public double Income { get; set; }
         public string Exchange { get; set; }
